# openapi.yaml
openapi: 3.0.3
info:
  title: Project Tracker API
  version: 1.0.0
  description: Cloud-agnostic API for managing users, customers, projects, and finance updates
servers:
  - url: https://api.example.com/v1
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List all customers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of customers
    post:
      tags: [Customers]
      summary: Create a new customer
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Customer:
      type: object
      required: [name, country, institution_type]
      properties:
        name:
          type: string
        country:
          type: string
        institution_type:
          type: string
        registration_number:
          type: string
        contacts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              phone:
                type: string
              mobile:
                type: string
              role:
                type: string
...